feat(main.py): create order correspondence script
create order correspondence script. It takes two PDFs, the first containing the order list and the
second containing the order list for the courier. Having identified the buyer's name, the product
purchased and the quantity in pdf1, a search carried out within pdf2 and the name of the product and
the quantity (if it is greater than 1) is inserted above the buyer's name in pdf2.