fix(main.py): manage products with the same name but different specifications
manage products with the same name but different specifications. Add orderDistinction function whose
task is to return the name of the product with the specification contained in round brackets setting
a maximum length. In the extract_product_details function I add controls aimed at manipulating the
name of the product that will be returned. In the add_product_name_to_pdf function I separate the
insert text relating to the text from that relating to the quantity so as to allow me to manage the
font size of the latter separately.