fix(Analyzer.py): Fix Homonymy bug cause in PDF1 buyer_name is long two row
Fix homonymy bug cause in PDF1 buyer_name is long two row and the script takes a wrong shipping_address. In extract_shipping_address function i take the text in the third row after "Spedire a:" and save it as shipping_address2 then in add_product_name_to_pdf when i control the homonymy i check if there is in the page buyer_name and shipping_addressif it is not i check if there is buyer_name and shipping_address2.